added support for multiple images
git-svn-id: file:///repos/festr/trunk@420 9051720a-062f-4424-aca9-25ff1f032159

+++ src/frontend/Commands/FrBaseCmd.h
@@ -25,7 +25,7 @@ public:
     FrPropMacro(FrMainController*, MainController);
 
 protected:
-    void UpdatePipelineForID(int id, int point);
+    void UpdatePipelineForID(unsigned long id, int point);
 };
 
 #endif // FR_BASE_CMD
