+++ src/frontend/Commands/FrLayerCmd.h
@@ -30,7 +30,7 @@ public:
     // Properties
     FrSetPropMacro(Actions, Action);    
     FrSetPropMacro(FrLayerDocObj*, DocObj);
-    void SetID(int id);
+    void SetID(unsigned long id);
 
 private:
     // Helper methods here
@@ -42,12 +42,12 @@ private:
     bool ChangeLayerParams();
     bool ChangeLayerColormap();
 
-    int GetActiveLayerID();
-    bool IsRoiLayer(int id);
+    unsigned long GetActiveLayerID();
+    bool IsRoiLayer(unsigned long id);
 
 private:
     bool m_isID;
-    int m_ID;
+    unsigned long m_ID;
 };
 
 #endif
